+++ src/types/widget.ts
@@ -1,0 +1,10 @@
+export type CTA = {
+  palette: 'primary' | 'secondary'
+  label: string
+}
+
+export type Image = {
+  src: string
+  hiddenHeaderContent?: string
+  hiddenHeaderLevel?: string
+}
