+++ static/admin/config.yml
@@ -14,13 +14,13 @@ collections:
     folder: "content/sections"
     create: true
     fields:
-      - {label: "Layout", name: "layout", widget: "select", options: ["hero", "left", "right", "vertical"], default: "left"}
-      - {label: "CTA", name: "cta", widget: cta, required: false}
-      - {label: "Title", name: "title", widget: "string"}
+      - {label: "Layout", name: layout, widget: "select", options: ["hero", "left", "right", "vertical"], default: "left"}
+      - {label: "CTA", name: cta, widget: cta, required: false}
+      - {label: "Title", name: title, widget: "string"}
       - {label: Image, name: image, widget: imageInHeader, required: false}
-      - {label: "Image position", name: "imagePosition", widget: "select", options: ["left", "right"], default: "left"}
-      - {label: "Contrast text", name: "contrastText", widget: "boolean", default: false}
-      - {label: "Text", name: "body", widget: "markdown", required: false}
+      - {label: "Image position", name: imagePosition, widget: "select", options: ["left", "right"], default: "left"}
+      - {label: "Contrast text", name: contrastText, widget: "boolean", default: false}
+      - {label: "Text", name: body, widget: "markdown", required: false}
       - {label: "Background Image", name: backgroundImage, widget: image, required: false}
 
   - name: "pages"
